+++ relay-common/src/time.rs
@@ -15,8 +15,22 @@ pub fn instant_to_date_time(instant: Instant) -> chrono::DateTime<chrono::Utc> {
     instant_to_system_time(instant).into()
 }
 
-/// A unix timestap (full seconds elapsed since 1970).
-#[derive(Clone, Copy, Eq, Ord, PartialEq, PartialOrd)]
+/// The conversion result of [`UnixTimestamp::to_instant`].
+///
+/// If the time is outside of what can be represented in an [`Instant`], this is `Past` or
+/// `Future`.
+#[derive(Clone, Copy, Debug)]
+pub enum MonotonicResult {
+    /// A time before the earliest representable `Instant`.
+    Past,
+    /// A representable `Instant`.
+    Instant(Instant),
+    /// A time after the latest representable `Instant`.
+    Future,
+}
+
+/// A unix timestamp (full seconds elapsed since 1970-01-01 00:00 UTC).
+#[derive(Clone, Copy, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub struct UnixTimestamp(u64);
 
 impl UnixTimestamp {
@@ -53,6 +67,41 @@ impl UnixTimestamp {
     /// Returns the number of seconds since the UNIX epoch start.
     pub fn as_secs(self) -> u64 {
         self.0
+    }
+
+    /// Converts the UNIX timestamp into an `Instant` based on the current system timestamp.
+    ///
+    /// Returns [`MonotonicResult::Instant`] if the timestamp can be represented. Otherwise, returns
+    /// [`MonotonicResult::Past`] or [`MonotonicResult::Future`].
+    ///
+    /// Note that the system time is subject to skew, so subsequent calls to `to_instant` may return
+    /// different values.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use std::time::{Duration, Instant};
+    /// use relay_common::{MonotonicResult, UnixTimestamp};
+    ///
+    /// let timestamp = UnixTimestamp::now();
+    /// if let MonotonicResult::Instant(instant) = timestamp.to_instant() {
+    ///    assert!((Instant::now() - instant) < Duration::from_millis(1));
+    /// }
+    /// ```
+    pub fn to_instant(self) -> MonotonicResult {
+        let now = Self::now();
+
+        if self > now {
+            match Instant::now().checked_add(self - now) {
+                Some(instant) => MonotonicResult::Instant(instant),
+                None => MonotonicResult::Future,
+            }
+        } else {
+            match Instant::now().checked_sub(now - self) {
+                Some(instant) => MonotonicResult::Instant(instant),
+                None => MonotonicResult::Past,
+            }
+        }
     }
 }